feat: 더미데이터 삭제 및 OverlapClass파일 분리
